test(plugins): split dependency-health snapshot cases out of the oversized status suite
